rewriting evaluation impl for attention

+++ dataset.py
@@ -67,7 +67,7 @@ class COCODataset(data.Dataset):
             return img, caption
         else:
             img_path = os.path.join(self.img_dir, self.coco.loadImgs(img_id)[0]['file_name'])
-            all_captions = [torch.LongTensor(self._numericalize_caption(ann['caption'])) for ann in self._get_annotations(img_id)]
+            all_captions = [self._numericalize_caption(ann['caption']) for ann in self._get_annotations(img_id)]
             return img, caption, img_path, all_captions
 
     def __len__(self):
